refactor(analysis): extract AnalysisVmInvoker shared helper class (#498)
Removes duplicate InvokeGetAnalysisFields/GetSearchQuery between
_AnalysisController and AnalysisWidgetDataSource. Adds consistent
null guards and TargetInvocationException unwrapping everywhere.
Closes #498
